// web/src/lib/types.ts
export interface TimelineData {
	date: Date;
	waterLevel: number;
	waterVolume: number;
	totalPumpedFlow: number;
	waterInflow: number;
	pumpFlow11: number;
	pumpFlow12: number;
	pumpFlow13: number;
	pumpFlow14: number;
	pumpFlow21: number;
	pumpFlow22: number;
	pumpFlow23: number;
	pumpFlow24: number;
	pumpPower11: number;
	pumpPower12: number;
	pumpPower13: number;
	pumpPower14: number;
	pumpPower21: number;
	pumpPower22: number;
	pumpPower23: number;
	pumpPower24: number;
	pumpFrequency11: number;
	pumpFrequency12: number;
	pumpFrequency13: number;
	pumpFrequency14: number;
	pumpFrequency21: number;
	pumpFrequency22: number;
	pumpFrequency23: number;
	pumpFrequency24: number;
	electricityPrice1High: number;
	electricityPrice2Normal: number;
}